Move the ModInfo to modinfo.py (#1660)
(cherry picked from commit 0848d3ab35733c6ad0659b8894754c9ddcb1bb7c)
