+++ insights/__init__.py
@@ -24,7 +24,7 @@ import sys
 import yaml
 
 from .core import Scannable, LogFileOutput, Parser, IniConfigFile  # noqa: F401
-from .core import FileListing, LegacyItemAccess, ModInfo, SysconfigOptions  # noqa: F401
+from .core import FileListing, LegacyItemAccess, SysconfigOptions  # noqa: F401
 from .core import YAMLParser, JSONParser, XMLParser, CommandParser  # noqa: F401
 from .core import AttributeDict  # noqa: F401
 from .core import Syslog  # noqa: F401
